SMILES param.
