+++ makefile
@@ -4,7 +4,7 @@ OUTDIR=output
 SDFDIR=sdf
 
 DIRS=$(OBJDIR) $(BINDIR) $(OUTDIR) $(SDFDIR)
-OBJS=$(OBJDIR)/misc.o $(OBJDIR)/point.o $(OBJDIR)/atom.o $(OBJDIR)/intera.o $(OBJDIR)/molecule.o $(OBJDIR)/aminoacid.o $(OBJDIR)/protein.o
+OBJS=$(OBJDIR)/misc.o $(OBJDIR)/point.o $(OBJDIR)/atom.o $(OBJDIR)/intera.o $(OBJDIR)/molecule.o $(OBJDIR)/aminoacid.o $(OBJDIR)/protein.o $(OBJDIR)/glom.o
 TESTS=test/point_test test/atom_test test/molecule_test test/pi_stack_test test/mol_assem_test test/amino_test test/aniso_test test/protein_test test/backbone_test
 APPS=$(BINDIR)/primarydock $(BINDIR)/pepteditor
 REPORTS=amino_report atom_report aniso_report point_report molecule_report mol_assem_report protein_report
@@ -61,6 +61,9 @@ $(OBJDIR)/aminoacid.o: src/classes/aminoacid.h src/classes/aminoacid.cpp $(OBJDI
 $(OBJDIR)/protein.o: src/classes/protein.h src/classes/protein.cpp $(OBJDIR)/aminoacid.o
 	$(CC) -c src/classes/protein.cpp -o $(OBJDIR)/protein.o $(CFLAGS)
 
+$(OBJDIR)/glom.o: src/classes/glom.h src/classes/glom.cpp $(OBJDIR)/protein.o
+	$(CC) -c src/classes/glom.cpp -o $(OBJDIR)/glom.o $(CFLAGS)
+
 test/point_test: src/point_test.cpp $(OBJDIR)/point.o
 	$(CC) src/point_test.cpp $(OBJDIR)/point.o $(OBJDIR)/misc.o -o test/point_test $(CFLAGS)
 
@@ -88,7 +91,7 @@ test/protein_test: src/protein_test.cpp $(OBJS) $(OBJDIR)/aminoacid.o $(OBJDIR)/
 test/backbone_test: src/backbone_test.cpp $(OBJS) $(OBJDIR)/aminoacid.o $(OBJDIR)/protein.o
 	$(CC) src/backbone_test.cpp $(OBJS) -o test/backbone_test $(CFLAGS)
 
-$(BINDIR)/primarydock: src/primarydock.cpp $(OBJS) $(OBJDIR)/aminoacid.o $(OBJDIR)/protein.o
+$(BINDIR)/primarydock: src/primarydock.cpp $(OBJS) $(OBJDIR)/aminoacid.o $(OBJDIR)/protein.o $(OBJDIR)/glom.o
 	$(CC) src/primarydock.cpp $(OBJS) -o $(BINDIR)/primarydock $(CFLAGS)
 
 $(BINDIR)/pepteditor: src/interpreter.cpp $(OBJS) $(OBJDIR)/aminoacid.o $(OBJDIR)/protein.o
