+++ primarydock.config
@@ -22,6 +22,19 @@ PROT pdbs/OR1/OR1A1.upright.pdb
 # Path to PDB or SDF of ligand (required):
 LIG sdf/citral.sdf
 
+# Optional SMILES input string. This will override the ligand SDF file.
+# It is not recommended to specify a completely different molecule from the SDF file.
+# This feature is intended for instances where e.g. a ligand should be docked in an
+# ionized state and the SDF contains a neutral uncharged molecule, or an internal
+# arrangement of conjugated bonds in the desired ligand differs from the SDF, or any
+# other variation not involving the addition or removal of heavy atoms, nor isomerization
+# of the molecule.
+# Note that if the config file specifies a SMILES string, and a --lig argument is used
+# on the command line, that argument will *not* override the SMILES string from the config
+# file. In this case, you will want to make sure to also specify a --smiles argument on the
+# command line.
+# SMILES CC(=CCCC(=CC=O)C)C
+
 # Center of binding pocket (required):
 # Can be ABS or RES; default is ABS if unspecified.
 # ABS means use literal XYZ coordinates.
