+++ test/testTAAR8.config
@@ -7,6 +7,7 @@ PROT pdbs/TAAR/TAAR8.upright.pdb
 
 # Path to PDB or SDF of ligand (required):
 LIG sdf/cadaverine.sdf
+SMILES [NH3+]CCCCC[NH3+]
 
 CEN RES 3.32 3.37 5.42 6.48 7.43
